CheckerMethod checks if company exists in DB via ID

+++ company/CompanyController.go
@@ -120,7 +120,39 @@ func DeleteCompany(c *gin.Context) {
 	if err != nil {
 		panic(err)
 	}
+
+	result := CheckIfCompanyExistsById(c)
+
+	if result == false {
+		c.IndentedJSON(http.StatusNotFound, gin.H{"Message": "company not found"})
+		return
+	} 
+
 	connection.Query("DELETE FROM companies WHERE id = ?", id)
 
 	c.IndentedJSON(http.StatusNoContent, gin.H{})
+}
+
+func CheckIfCompanyExistsById(c *gin.Context) bool {
+	connection, err := services.GetDbConnection(c)
+
+	if err != nil {
+		panic(err)
+	}
+	id, err := hex.DecodeString(c.Param("id"))
+	if err != nil {
+		panic(err)
+	}
+
+	item, err := connection.Query("SELECT * FROM companies WHERE id = ?", id)
+
+	if err != nil {
+		panic(err)
+	}
+
+	if item.Next() {
+		return true
+	} else {
+		return false
+	}
 }
